Modularize main.js -  dataLoader.js, iconManager.js, optionManager.js
